Basic Example and Experiments Working
