--- Examples/Others/RLLib_Training.py
@@ -1,115 +1,0 @@
-"""Uses Ray's RLlib to train agents to play Pistonball.
-
-Author: Rohan (https://github.com/Rohan138)
-"""
-
-import os
-
-import ray
-import supersuit as ss
-from ray import tune
-from ray.rllib.algorithms.ppo import PPOConfig
-from ray.rllib.env.wrappers.pettingzoo_env import ParallelPettingZooEnv
-from ray.rllib.models import ModelCatalog
-from ray.rllib.models.torch.torch_modelv2 import TorchModelV2
-from ray.tune.registry import register_env
-from torch import nn
-
-from pettingzoo.mpe import simple_spread_v3
-
-
-class CNNModelV2(TorchModelV2, nn.Module):
-    def __init__(self, obs_space, act_space, num_outputs, *args, **kwargs):
-        TorchModelV2.__init__(self, obs_space, act_space, num_outputs, *args, **kwargs)
-        nn.Module.__init__(self)
-        self.model = nn.Sequential(
-            nn.Conv2d(3, 32, [2, 2]),
-            nn.ReLU(),
-            nn.Conv2d(32, 64, [2, 2]),
-            nn.ReLU(),
-            nn.Conv2d(64, 64, [2, 2]),
-            nn.ReLU(),
-            nn.Flatten(),
-            (nn.Linear(1024, 64)),
-            nn.ReLU(),
-        )
-        self.policy_fn = nn.Linear(64, 5)
-        self.value_fn = nn.Linear(64, 1)
-     
-    def forward(self, obs, state=None, info=None):
-        # Permute the input dimensions to (batch_size, channels, height, width)
-        obs_tensor = torch.tensor(obs, dtype=torch.float32).to(self.device)
-        obs_permuted = obs_tensor.permute(0, 3, 1, 2)        
-        model_out = self.model(obs_permuted)
-        self._value_out = self.value_fn(model_out)
-        return self.policy_fn(model_out), state
-
-    def value_function(self):
-        return self._value_out.flatten()
-
-
-def env_creator(args):
-    env = simple_spread_v3.parallel_env(
-        # n_pistons=20,
-        # time_penalty=-0.1,
-        # continuous=True,
-        # random_drop=True,
-        # random_rotate=True,
-        # ball_mass=0.75,
-        # ball_friction=0.3,
-        # ball_elasticity=1.5,
-        # max_cycles=125,
-        continuous_actions=False
-    )
-    # env = ss.color_reduction_v0(env, mode="B")
-    # env = ss.dtype_v0(env, "float32")
-    # env = ss.resize_v1(env, x_size=10, y_size=10)
-    # env = ss.normalize_obs_v0(env, env_min=0, env_max=1)
-    # env = ss.frame_stack_v1(env, 3)
-    return env
-
-
-if __name__ == "__main__":
-    ray.init()
-
-    env_name = "single_spread_v3"
-    
-    env = env_creator({})
-
-    register_env(env_name, lambda config: ParallelPettingZooEnv(env_creator(config)))
-    # ModelCatalog.register_custom_model("CNNModelV2", CNNModelV2)
-    
-    config = (
-        PPOConfig()
-        .environment(env=env_name, clip_actions=True)
-        .rollouts(num_rollout_workers=4, rollout_fragment_length=128)
-        .training(
-            train_batch_size=512,
-            lr=1e-4,
-            gamma=0.99,
-            lambda_=0.9,
-            use_gae=True,
-            # clip_param=1.0,
-            grad_clip=None,
-            entropy_coeff=0.0,
-            vf_loss_coeff=0.25,
-            sgd_minibatch_size=64,
-            num_sgd_iter=20,
-        )
-        .debugging(log_level="ERROR")
-        .framework(framework="torch")
-        .resources(num_gpus=int(os.environ.get("RLLIB_NUM_GPUS", "1")))  
-        .multi_agent(
-            policies=env.possible_agents,
-            policy_mapping_fn=(lambda agent_id, *args, **kwargs: agent_id),
-        )      
-    )
-
-    tune.run(
-        "PPO",
-        name="PPO",
-        stop={"timesteps_total": 5000000 if not os.environ.get("CI") else 50000},
-        checkpoint_freq=20,
-        local_dir="~/ray_results/" + env_name,
-        config=config.to_dict(),
-    )
